Assignment2/requirements.txt
numpy==1.19.2
snap-stanford==5.0.0
